refactor(frontend): split playtest active rooms
